Split Rust caps and database crates
