+++ Makefile.am
@@ -11,7 +11,10 @@ EXTRA_DIST = config.rpath intltool-extract.in intltool-merge.in intltool-update.
 	crates/ddccontrol-caps/fixtures/db_patch_add_remove.caps \
 	crates/ddccontrol-caps/fixtures/fallback_crt.caps \
 	crates/ddccontrol-caps/fixtures/nested_values.caps \
-	crates/ddccontrol-caps/src/lib.rs
+	crates/ddccontrol-caps/src/lib.rs \
+	crates/ddccontrol-db/Cargo.toml \
+	crates/ddccontrol-db/README.md \
+	crates/ddccontrol-db/src/lib.rs
 DISTCLEANFILES = intltool-extract intltool-merge intltool-update
 DISTCHECK_CONFIGURE_FLAGS = --enable-doc
