+++ crates/ddccontrol-db/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "ddccontrol-db"
+version = "0.1.0"
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Rust monitor database XML parser and C ABI bridge for ddccontrol"
+
+[lib]
+crate-type = ["staticlib"]
+
+[dependencies]
+ddccontrol-caps = { path = "../ddccontrol-caps" }
+roxmltree = "0.20"
