+++ crates/ddccontrol-db/README.md
@@ -1,0 +1,6 @@
+# ddccontrol-db
+
+`ddccontrol-db` parses the ddccontrol XML monitor database and exports the C ABI
+used by `libddccontrol`.
+
+CAPS string parsing lives in the sibling `ddccontrol-caps` crate.
