+++ src/lib/Makefile.am
@@ -15,8 +15,8 @@ EXTRA_DIST = $(modules_DATA)
 
 lib_LTLIBRARIES = libddccontrol.la
 
-RUST_CAPS_LIB = $(top_builddir)/target/release/libddccontrol_caps.a
-RUST_CAPS_DEPS = \
+RUST_DB_LIB = $(top_builddir)/target/release/libddccontrol_db.a
+RUST_DB_DEPS = \
 	$(top_srcdir)/Cargo.toml \
 	$(top_srcdir)/Cargo.lock \
 	$(top_srcdir)/crates/ddccontrol-caps/Cargo.toml \
@@ -25,17 +25,19 @@ RUST_CAPS_DEPS = \
 	$(top_srcdir)/crates/ddccontrol-caps/fixtures/compact_vcp.caps \
 	$(top_srcdir)/crates/ddccontrol-caps/fixtures/db_patch_add_remove.caps \
 	$(top_srcdir)/crates/ddccontrol-caps/fixtures/fallback_crt.caps \
-	$(top_srcdir)/crates/ddccontrol-caps/fixtures/nested_values.caps
+	$(top_srcdir)/crates/ddccontrol-caps/fixtures/nested_values.caps \
+	$(top_srcdir)/crates/ddccontrol-db/Cargo.toml \
+	$(top_srcdir)/crates/ddccontrol-db/src/lib.rs
 
-$(RUST_CAPS_LIB): $(RUST_CAPS_DEPS)
-	cd $(top_srcdir) && DDCONTROL_DATADIR="$(datadir)/ddccontrol-db" CARGO_TARGET_DIR=$(abs_top_builddir)/target $(CARGO) build --release -p ddccontrol-caps
+$(RUST_DB_LIB): $(RUST_DB_DEPS)
+	cd $(top_srcdir) && DDCONTROL_DATADIR="$(datadir)/ddccontrol-db" CARGO_TARGET_DIR=$(abs_top_builddir)/target $(CARGO) build --release -p ddccontrol-db
 
 libddccontrol_la_SOURCES = ddcci.c ddcci.h monitor_db.h \
    i2c-dev.h conf.c conf.h internal.h \
    monitor_db_internal.h
 
 libddccontrol_la_LDFLAGS = -version-info 1:0:0
-libddccontrol_la_LIBADD = $(RUST_CAPS_LIB) $(LIBXML2_LIBS)
+libddccontrol_la_LIBADD = $(RUST_DB_LIB) $(LIBXML2_LIBS)
 
 check_PROGRAMS = test_issueurl test_parse_caps test_edid
 TESTS = $(check_PROGRAMS)
